BJS2-84236-leaderboard - написал docker-compose для redis insight - изменил Long в RedisTemplate<String, Long> на Object - в репозитории написал upsert запрос - реализовал логику в LeaderboardService

+++ src/main/java/school/faang/user_service/rating_service/entity/DefaultScore.java
@@ -45,7 +45,7 @@ public class DefaultScore {
     private ActionType actionType;
 
     @Column(name = "base_points", nullable = false)
-    private int basePoints;
+    private Double basePoints;
 
     @Column(name = "is_active", nullable = false)
     private Boolean isActive;
